refactor: add abstract class for io operations
