--- libs/io_interface.py
@@ -1,1 +1,0 @@
-#! todo: implement interface for writer and reader
